+++ api_app/analyzers_manager/urls.py
@@ -11,6 +11,7 @@ router = routers.DefaultRouter(trailing_slash=False)
 router.register(
     r"jobs/(?P<job_id>\d+)/analyzer/(?P<report_id>\w+)",
     AnalyzerActionViewSet,
+    basename="analyzerreport",
 )
 router.register(r"analyzer", AnalyzerConfigAPI, basename="analyzer")