+++ api_app/playbooks_manager/models.py
@@ -12,7 +12,7 @@ from api_app.validators import plugin_name_validator, validate_runtime_configura
 
 class PlaybookConfig(models.Model):
     name = models.CharField(
-        max_length=30,
+        max_length=100,
         null=False,
         unique=True,
         primary_key=True,
@@ -37,3 +37,6 @@ class PlaybookConfig(models.Model):
         null=False,
         validators=[validate_runtime_configuration],
     )
+
+    class Meta:
+        ordering = ["name", "disabled"]