+++ api_app/visualizers_manager/views.py
@@ -3,7 +3,8 @@
 
 import logging
 
-from api_app.core.views import AbstractConfigAPI
+from api_app.core.views import AbstractConfigAPI, PluginActionViewSet
+from api_app.visualizers_manager.models import VisualizerReport
 from api_app.visualizers_manager.serializers import VisualizerConfigSerializer
 
 logger = logging.getLogger(__name__)
@@ -16,3 +17,16 @@ __all__ = [
 
 class VisualizerConfigAPI(AbstractConfigAPI):
     serializer_class = VisualizerConfigSerializer
+
+
+class VisualizerActionViewSet(PluginActionViewSet):
+    @classmethod
+    @property
+    def report_model(cls):
+        return VisualizerReport
+
+    def perform_retry(self, report: VisualizerReport):
+        signature = report.config.get_signature(
+            report.job,
+        )
+        signature.apply_async()